requirements.txt
discord.py>=2.3.2
anthropic>=0.18.1